requirements.txt
# glotzer tools
freud-analysis==2.7.0
gsd==2.5.0
garnett==0.7.1
pythia-learn==0.3.0
rowan==1.3.0.post1
coxeter==0.6.1
plato-draw==1.12.0
signac==1.7.0
signac-flow==0.16.0

# commonly used packages
h5py==3.5.0
jupyter==1.0.0; platform_machine != 'ppc64le'
matplotlib==3.4.3; platform_machine != 'ppc64le'
notebook==6.4.5; platform_machine != 'ppc64le'
opencv-python==4.5.4.58
pandas==1.3.4; platform_machine != 'ppc64le'
Pillow==8.4.0
pyqt5==5.15.6; platform_machine != 'ppc64le'
pyyaml==6.0
scikit-learn==1.0.1
scikit-image==0.18.3; platform_machine != 'ppc64le'
scipy==1.7.1
statsmodels==0.13.0

# test requirements
hypothesis==6.24.1
pytest==6.2.5
miniball==1.1.0
sympy==1.9
tables==3.6.1
